Add types to GeneViewer
