+++ app/views/ReportView/components/SmallMutations/index.tsx
@@ -8,8 +8,8 @@ import snackbar from '@/services/SnackbarUtils';
 import DataTable from '@/components/DataTable';
 import ReportContext from '@/context/ReportContext';
 import { WithLoadingInjectedProps } from '@/hoc/WithLoading';
+import { SmallMutationType } from '@/common';
 import { columnDefs } from './columnDefs';
-import MutationType from './types';
 
 import './index.scss';
 
@@ -36,7 +36,7 @@ const SmallMutations = ({
   setIsLoading,
 }: SmallMutationsProps): JSX.Element => {
   const { report } = useContext(ReportContext);
-  const [smallMutations, setSmallMutations] = useState<MutationType[]>([]);
+  const [smallMutations, setSmallMutations] = useState<SmallMutationType[]>([]);
   const [groupedSmallMutations, setGroupedSmallMutations] = useState({
     therapeutic: [],
     nostic: [],
